requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
slack-sdk==3.23.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2
schedule==1.2.0
pytz==2023.3
requests==2.31.0
